+++ inst/tinytest/test-messages.R
@@ -12,7 +12,7 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 6, info = "Print step messages")
+expect_equal(length(get_message_stack()), 6, info = "Print normal messages utf8")
 expect_message(print_stack_as_messages("NOTE"),    "This is a note.",    info = "Print normal messages utf8")
 expect_warning(print_stack_as_messages("WARNING"), "This is a warning.", info = "Print normal messages utf8")
 expect_error(print_stack_as_messages("ERROR"),     "This is an error.",  info = "Print normal messages utf8")
@@ -31,7 +31,7 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 6, info = "Print step messages")
+expect_equal(length(get_message_stack()), 6, info = "Print normal messages non-utf8")
 expect_message(print_stack_as_messages("NOTE"),    "This is a note.",    info = "Print normal messages non-utf8")
 expect_warning(print_stack_as_messages("WARNING"), "This is a warning.", info = "Print normal messages non-utf8")
 expect_error(print_stack_as_messages("ERROR"),     "This is an error.",  info = "Print normal messages non-utf8")
@@ -50,7 +50,7 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 6, info = "Print step messages")
+expect_equal(length(get_message_stack()), 6, info = "Print step messages utf8")
 expect_message(print_stack_as_messages("GREY"),  "Grey step.",  info = "Print step messages utf8")
 expect_message(print_stack_as_messages("MAJOR"), "Major step.", info = "Print step messages utf8")
 expect_message(print_stack_as_messages("MINOR"), "Minor step.", info = "Print step messages utf8")
@@ -68,7 +68,7 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 6, info = "Print step messages")
+expect_equal(length(get_message_stack()), 6, info = "Print step messages non-utf8")
 expect_message(print_stack_as_messages("GREY"),  "Grey step.",  info = "Print step messages non-utf8")
 expect_message(print_stack_as_messages("MAJOR"), "Major step.", info = "Print step messages non-utf8")
 expect_message(print_stack_as_messages("MINOR"), "Minor step.", info = "Print step messages non-utf8")
@@ -89,7 +89,7 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 4, info = "Print step messages")
+expect_equal(length(get_message_stack()), 4, info = "Print custom messages utf8")
 expect_message(print_stack_as_messages("UNICORN"), "This is a unicorn message.", info = "Print custom messages utf8")
 expect_message(print_stack_as_messages("UNICORN"), "This is a unicorn step.",    info = "Print custom messages utf8")
 
@@ -106,9 +106,27 @@ test_func <- function(){
 
 test_func()
 
-expect_equal(length(get_message_stack()), 4, info = "Print step messages")
+expect_equal(length(get_message_stack()), 4, info = "Print custom messages non-utf8")
 expect_message(print_stack_as_messages("UNICORN"), "This is a unicorn message.", info = "Print custom messages non-utf8")
 expect_message(print_stack_as_messages("UNICORN"), "This is a unicorn step.",    info = "Print custom messages non-utf8")
+
+
+# Print in place messages
+test_func <- function(){
+    print_start_message()
+
+    for (i in seq_len(10)){
+        print_step("Minor", "This is in place step [i] of 10", i = i, in_place = TRUE)
+    }
+
+    print_closing()
+}
+
+test_func()
+
+expect_equal(length(get_message_stack()), 12, info = "Print in place messages")
+expect_message(print_stack_as_messages("MINOR"), "This is in place step 1 of 10",  info = "Print custom messages non-utf8")
+expect_message(print_stack_as_messages("MINOR"), "This is in place step 10 of 10", info = "Print custom messages non-utf8")
 
 
 # Print messages without start and end only return a message stack of length 1